cli/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.7
fastapi
uvicorn
sqlalchemy
pydantic